Moving shuffling up a level
